fix(observability): seal retained logger facade
